requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.2.0
celery==5.3.6
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.7
dj-inmemorystorage==2.1.0
Django==5.0
django-debug-toolbar==4.2.0
django-debug-toolbar-force==0.2
django-filter==23.5
django-nine==0.2.7
django-phonenumber-field==7.2.0
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
flower==2.0.1
humanize==4.9.0
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kombu==5.3.4
Markdown==3.5.1
mock==5.1.0
model-bakery==1.17.0
packaging==23.2
phonenumberslite==8.13.26
Pillow==10.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psycopg2
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.32.0
rpds-py==0.13.2
six==1.16.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
tornado==6.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.12
